+++ etcd_dynamic_config/tests/__init__.py
@@ -1,1 +1,1 @@
-# Tests for etcd-config-provider
+# Tests for etcd-dynamic-config